refacto(organization): out of legacy (2nd)
