+++ src/bootstrap.ts
@@ -20,6 +20,8 @@ import { NotFound } from "./not-found";
 import { proxy } from "./proxy/route";
 import welcome_router from "./welcome/route";
 
+//
+
 const app = new Hono<Csp_Context>()
   .use("*", logger())
   .use("*", csp_headers())
